Add missing SQL types to PreferenceOrder and optimize lookups
Core fixes:
- Added byte, short, long, Guid, and byte[] to PreferenceOrder
- These types were supported by deciders but not in PreferenceOrder
- This caused SetInitialTypeHint() to ignore them, breaking type hints

Performance improvements:
- Added PreferenceOrderIndex (FrozenDictionary<Type, int>) for O(1) lookups
- Replaced PreferenceOrder.IndexOf() with PreferenceOrderIndex[] in PooledBuilder
- Maintains same ordering semantics with better performance

Test updates:
- Updated tests to use System.Drawing.Color as truly unsupported type
- Guid is now supported, so tests expecting it to fail needed updating
- Added PreferenceOrderIndex_MatchesPreferenceOrder test
- Added PreferenceOrder_ContainsAllCommonSqlTypes test
- Use Is.EqualTo().AsCollection for cleaner sequence assertions

Type ordering rationale:
- bool, int, decimal, long (common types first)
- TimeSpan, DateTime (temporal types)
- byte, short (less common, more specific integer types)
- Guid, byte[] (special types, never guessed from strings)
- string (fallback)

All 379 tests passing.

+++ TypeGuesser/Deciders/IntTypeDecider.cs
@@ -10,7 +10,7 @@ namespace TypeGuesser.Deciders;
 /// Creates a new instance for recognizing whole numbers in string values
 /// </remarks>
 /// <param name="culture"></param>
-public sealed class IntTypeDecider(CultureInfo culture) : DecideTypesForStrings<int>(culture,TypeCompatibilityGroup.Numerical, typeof(short) , typeof(int), typeof(byte))
+public sealed class IntTypeDecider(CultureInfo culture) : DecideTypesForStrings<int>(culture,TypeCompatibilityGroup.Numerical, typeof(byte), typeof(short), typeof(int))
 {
     /// <inheritdoc/>
     protected override IDecideTypesForStrings CloneImpl(CultureInfo newCulture)
